requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Database & ORM
sqlalchemy
psycopg2-binary
alembic

# Data Analysis & File Handling
polars
python-multipart
openpyxl

# Authentication & Security
passlib[bcrypt]
python-jose[cryptography]

# Presentation Generation
python-pptx

# Environment Configuration
python-dotenv
